requirements.txt
# Required:
## Data structure
numpy
pandas
pydantic
## ML lib
scikit-learn
## Web interface
fastapi
pyyaml  # server configurations
python-multipart  # file service
uvicorn
gunicorn
# Optional:
pytest
coverage